rename project name + add a few packages to poetry
