--- src/bluecoins_cli/__main__.py
@@ -1,4 +1,0 @@
-from bluecoins_cli.cli import root
-
-if __name__ == '__main__':
-    root(prog_name='bluecoins-cli', standalone_mode=False)
