+++ src/cluecoins/__main__.py
@@ -1,0 +1,4 @@
+from cluecoins.cli import root
+
+if __name__ == '__main__':
+    root(prog_name='cluecoins', standalone_mode=False)
